# webserver/app/static/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "description": "Federated Node",
    "title": "Federated Node",
    "version": "0.0.1"
  },
  "servers":[
    {
      "url": "/",
      "description": "Live service"
    },
    {
      "url": "http://127.0.0.1:5000",
      "description": "Local server"
    }
  ],
...
  "paths":{
    "/datasets":{
      "get": {
        "operationId": "getDataset",
        "tags": ["Datasets"],
        "summary": "List all datasets available to the user",
        "responses": {
          "200":{
            "$ref": "#/components/responses/DatasetList"
          },
          "400":{
            "$ref": "#/components/responses/InvalidBody"
          },
          "401":{
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403":{
            "$ref": "#/components/responses/Unauthorized"
          },
          "500":{
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "post":{
        "operationId": "postDataset",
        "tags": ["Datasets"],
        "summary": "Create a new dataset that the Federated Node can fetch data from",
        "requestBody": {
          "content": {
            "application/json":{
              "schema":{
                "$ref": "#/components/schemas/DatasetPostBody"
              }
            }
          }
        },
        "responses":{
          "201":{
            "$ref": "#/components/responses/DatasetPost"
          },
          "400":{
            "$ref": "#/components/responses/InvalidBody"
          },
          "401":{
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403":{
            "$ref": "#/components/responses/Unauthorized"
          },
          "404":{
            "$ref": "#/components/responses/NotFound"
          },
          "500":{
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/datasets/{id}": {
      "get": {
        "operationId": "getDatasetById",
        "parameters": [
          {
            "$ref": "#/components/parameters/datasetIdPath"
          },
          {
            "$ref": "#/components/parameters/projectNamePath"
          }
        ],
        "tags": ["Datasets"],
        "summary": "Get a datasets by id if available to the user",
        "responses": {
          "200":{
            "$ref": "#/components/responses/DatasetById"
          },
          "401":{
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403":{
            "$ref": "#/components/responses/Unauthorized"
          },
          "404":{
            "$ref": "#/components/responses/NotFound"
          },
          "500":{
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "patch":{
        "operationId": "editDataset",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema":{"type": "integer"},
            "required": true
          }
        ],
        "tags": ["Datasets"],
        "summary": "Edit a dataset information. Any of the fields provided on the POST or returned by the GET on this same endpoint, is a valid body",
        "requestBody": {
          "content": {
            "application/json":{
              "schema":{
                "$ref": "#/components/schemas/DatasetPostBody"
              }
            }
          }
        },
        "responses": {
          "204": {
            "$ref": "#/components/responses/DatasetPost"
          },
          "400":{
            "$ref": "#/components/responses/InvalidBody"
          },
          "401":{
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403":{
            "$ref": "#/components/responses/Unauthorized"
          },
          "404":{
            "$ref": "#/components/responses/NotFound"
          },
          "500":{
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "delete":{
        "operationId": "deleteDataset",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema":{"type": "integer"},
            "required": true
          }
        ],
        "tags": ["Datasets"],
        "summary": "Deletes a dataset mapping with its kubernetes secrets.",
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContent"
          },
          "400":{
            "$ref": "#/components/responses/InvalidBody"
          },
          "401":{
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403":{
            "$ref": "#/components/responses/Unauthorized"
          },
          "404":{
            "$ref": "#/components/responses/NotFound"
          },
          "500":{
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/datasets/{id}/catalogue": {
      "get": {
        "operationId": "getDatasetCatalogue",
        "parameters": [
          {
            "$ref": "#/components/parameters/datasetIdPath"
          },
          {
            "$ref": "#/components/parameters/projectNamePath"
          }
        ],
        "tags": ["Datasets"],
        "summary": "Get the catalogue in a given dataset",
        "responses": {
          "200":{
            "$ref": "#/components/responses/DatasetCatalogue"
          },
          "401":{
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403":{
            "$ref": "#/components/responses/Unauthorized"
          },
          "404":{
            "$ref": "#/components/responses/NotFound"
          },
          "500":{
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/datasets/{id}/dictionaries": {
      "get": {
        "operationId": "getDatasetDictionaries",
        "parameters": [
          {
            "$ref": "#/components/parameters/datasetIdPath"
          },
          {
            "$ref": "#/components/parameters/projectNamePath"
          }
        ],
        "tags": ["Datasets"],
        "summary": "Get the dictionaries in a given dataset",
        "responses": {
          "200":{
            "$ref": "#/components/responses/DatasetDictionary"
          },
          "401":{
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403":{
            "$ref": "#/components/responses/Unauthorized"
          },
          "404":{
            "$ref": "#/components/responses/NotFound"
          },
          "500":{
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/datasets/{id}/dictionaries/{table_name}": {
      "get": {
        "operationId": "getDatasetDictionariesTable",
        "parameters": [
          {
            "$ref": "#/components/parameters/datasetIdPath"
          },
          {
            "$ref": "#/components/parameters/dictionaryTablePath"
          },
          {
            "$ref": "#/components/parameters/projectNamePath"
          }
        ],
        "tags": ["Datasets"],
        "summary": "Get the dictionary for a given table in a given dataset",
        "responses": {
          "200":{
            "$ref": "#/components/responses/DatasetDictionaryTable"
          },
          "401":{
            "$ref": "#/components/responses/Unauthenticated"
          },
          "404":{
            "$ref": "#/components/responses/NotFound"
          },
          "403":{
            "$ref": "#/components/responses/Unauthorized"
          },
          "500":{
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/datasets/{name}": {
      "get": {
        "operationId": "getDatasetByName",
        "parameters": [
          {
            "$ref": "#/components/parameters/datasetNamePath"
          },
          {
            "$ref": "#/components/parameters/projectNamePath"
          }
        ],
        "tags": ["Datasets"],
        "summary": "Get a datasets by name if available to the user",
        "responses": {
          "200":{
            "$ref": "#/components/responses/DatasetById"
          },
          "401":{
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403":{
            "$ref": "#/components/responses/Unauthorized"
          },
          "404":{
            "$ref": "#/components/responses/NotFound"
          },
          "500":{
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "patch": {
        "operationId": "editDataset",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema":{"type": "integer"},
            "required": true
          }
        ],
        "tags": ["Datasets"],
        "summary": "Edit a dataset information. Any of the fields provided on the POST or returned by the GET on this same endpoint, is a valid body",
        "requestBody": {
          "content": {
            "application/json":{
              "schema":{
                "$ref": "#/components/schemas/DatasetPostBody"
              }
            }
          }
        },
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContent"
          },
          "400":{
            "$ref": "#/components/responses/InvalidBody"
          },
          "401":{
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403":{
            "$ref": "#/components/responses/Unauthorized"
          },
          "404":{
            "$ref": "#/components/responses/NotFound"
          },
          "500":{
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "delete": {
        "operationId": "deleteDataset",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema":{"type": "integer"},
            "required": true
          }
        ],
        "tags": ["Datasets"],
        "summary": "Deletes a dataset mapping with its kubernetes secrets.",
        "responses": {
          "204": {
            "$ref": "#/components/responses/DatasetPost"
          },
          "400":{
            "$ref": "#/components/responses/InvalidBody"
          },
          "401":{
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403":{
            "$ref": "#/components/responses/Unauthorized"
          },
          "404":{
            "$ref": "#/components/responses/NotFound"
          },
          "500":{
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/datasets/{name}/catalogue": {
      "get": {
        "operationId": "getDatasetCatalogue",
        "parameters": [
          {
            "$ref": "#/components/parameters/datasetNamePath"
          },
          {
            "$ref": "#/components/parameters/projectNamePath"
          }
        ],
        "tags": ["Datasets"],
        "summary": "Get the catalogue in a given dataset",
        "responses": {
          "200":{
            "$ref": "#/components/responses/DatasetCatalogue"
          },
          "401":{
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403":{
            "$ref": "#/components/responses/Unauthorized"
          },
          "404":{
            "$ref": "#/components/responses/NotFound"
          },
          "500":{
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/datasets/{name}/dictionaries": {
      "get": {
        "operationId": "getDatasetDictionaries",
        "parameters": [
          {
            "$ref": "#/components/parameters/datasetNamePath"
          },
          {
            "$ref": "#/components/parameters/projectNamePath"
          }
        ],
        "tags": ["Datasets"],
        "summary": "Get the dictionaries in a given dataset",
        "responses": {
          "200":{
            "$ref": "#/components/responses/DatasetDictionary"
          },
          "401":{
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403":{
            "$ref": "#/components/responses/Unauthorized"
          },
          "404":{
            "$ref": "#/components/responses/NotFound"
          },
          "500":{
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/datasets/{name}/dictionaries/{table_name}": {
      "get": {
        "operationId": "getDatasetDictionariesTable",
        "parameters": [
          {
            "$ref": "#/components/parameters/datasetNamePath"
          },
          {
            "$ref": "#/components/parameters/dictionaryTablePath"
          },
          {
            "$ref": "#/components/parameters/projectNamePath"
          }
        ],
        "tags": ["Datasets"],
        "summary": "Get the dictionary for a given table in a given dataset",
        "responses": {
          "200":{
            "$ref": "#/components/responses/DatasetDictionaryTable"
          },
          "401":{
            "$ref": "#/components/responses/Unauthenticated"
          },
          "404":{
            "$ref": "#/components/responses/NotFound"
          },
          "403":{
            "$ref": "#/components/responses/Unauthorized"
          },
          "500":{
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/datasets/token_transfer": {
      "post": {
        "operationId": "post_transfer_token",
        "tags": ["Tokens"],
        "summary": "Endpoint to create a user's permission for a dataset, and returns a token for it",
        "requestBody": {
          "content": {
            "application/json":{
              "schema":{
                "$ref": "#/components/schemas/TokenPostBody"
              }
            }
          }
        },
        "responses":{
          "201":{
            "$ref": "#/components/responses/TokenResponse"
          },
          "400":{
            "$ref": "#/components/responses/InvalidBody"
          },
          "401":{
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403":{
            "$ref": "#/components/responses/Unauthorized"
          },
          "500":{
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/datasets/selection/beacon": {
      "post": {
        "operationId": "select_beacon",
        "tags": ["Datasets"],
        "summary": "Endpoint to check if a given query is valid",
        "requestBody": {
          "content": {
            "application/json":{
              "schema":{
                "$ref": "#/components/schemas/BeaconPostBody"
              }
            }
          }
        },
        "responses":{
          "201":{
            "$ref": "#/components/responses/SelectBeaconPost"
          },
          "401":{
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403":{
            "$ref": "#/components/responses/Unauthorized"
          },
          "500":{
            "$ref": "#/components/responses/DBConnectionFailed"
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "operationId": "get_all_users",
        "tags": ["Users"],
        "summary": "List of all of the users",
        "responses": {
          "200": {
            "$ref": "#/components/responses/UserListResponse"
          },
          "400": {
            "$ref": "#/components/responses/InvalidBody"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500":{
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "post": {
        "operationId": "create_user",
        "tags": ["Users"],
        "summary": "Create a new user",
        "responses": {
          "201": {
            "$ref": "#/components/responses/UserEntry"
          },
          "400": {
            "$ref": "#/components/responses/InvalidBody"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500":{
            "$ref": "#/components/responses/InternalError"
          }
        },
        "requestBody": {
          "content": {
            "application/json":{
              "schema":{
                "$ref": "#/components/schemas/CreateUserBody"
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components":{
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "datasetIdPath": {
        "in": "path",
        "name": "dataset_id",
        "schema":{"type": "integer"},
        "required": true
      },
      "datasetNamePath": {
        "in": "path",
        "name": "dataset_name",
        "schema":{"type": "string"},
        "required": true,
        "description": "Dataset name, the same set from the POST /dataset, or fetched from the dataset list"
      },
      "dictionaryTablePath": {
        "in": "path",
        "name": "table_name",
        "schema":{"type": "string"},
        "required": true
      },
      "projectNamePath": {
        "in": "header",
        "name": "project_name",
        "schema":{"type": "string"},
        "required": false,
        "description": "Used by non admin users. The value should match the project decalred on the roken_transfer endpoint"
      },
...
    },
    "responses": {
      "DatasetPost": {
        "description": "Successfully added",
        "content": {
          "application/json":{
            "schema":{
              "type": "object",
              "properties": {
                "dataset_id": {"type": "integer"}
              }
            }
          }
        }
      },
      "DatasetById": {
        "description": "Successfully added",
        "content": {
          "application/json":{
            "schema":{
              "$ref": "#/components/schemas/DatasetById"
            }
          }
        }
      },
      "DatasetList": {
        "description": "Successfully added",
        "content": {
          "application/json":{
            "schema":{
              "type": "object",
              "properties": {
                "datasets":{
                  "type": "array",
                  "items":{
                    "$ref": "#/components/schemas/DatasetById"
                  }
                }
              }
            }
          }
        }
      },
      "DatasetCatalogue": {
        "description": "Successfully added",
        "content": {
          "application/json":{
            "schema":{
              "type": "array",
              "items":{
                "$ref": "#/components/schemas/DatasetCatalogue"
              }
            }
          }
        }
      },
      "DatasetDictionary": {
        "description": "Successfully added",
        "content": {
          "application/json":{
            "schema":{
              "type": "array",
              "items":{
                "$ref": "#/components/schemas/DatasetDictionary"
              }
            }
          }
        }
      },
      "DatasetDictionaryTable": {
        "description": "Successfully added",
        "content": {
          "application/json":{
            "schema":{
              "$ref": "#/components/schemas/DatasetDictionary"
            }
          }
        }
      },
...
      "TokenResponse": {
        "description": "Token created",
        "content": {
          "application/json":{
            "schema":{
              "type": "object",
              "properties": {
                "token": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "SelectBeaconPost": {
        "description": "Successful beacon. Query is correct",
        "content": {
          "application/json":{
            "schema":{
              "type": "object",
              "properties": {
                "query": {"type": "string", "example": "SELECT * FROM patients;"},
                "result": {"type": "string", "example": "Ok"}
              }
            }
          }
        }
      },
...
      "NoContent": {
        "description": "No content response"
      },
...
      "UserEntry":{
        "description": "User Definition",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "email": {"type": "string", "example": "user@email.com"},
                "username": {"type": "string", "example": "user"},
                "info": {"type": "string", "example": "The user should change the temp password at https://localhost/users/reset-password"},
                "tempPassword": {"type": "string"}
              }
            }
          }
        }
      },
      "UserListResponse": {
        "description": "List of users",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items":{
                "$ref": "#/components/schemas/UserEntry"
              }
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Not found."
      },
      "InvalidBody": {
        "description": "The request body cannot be parsed or is missing required values."
      },
      "Unauthenticated": {
        "description": "Unauthenticated."
      },
      "Unauthorized": {
        "description": "Unauthorized."
      },
      "InternalError": {
        "description": "An internal error occurred."
      },
      "DBConnectionFailed": {
        "description": "Could not connect to the database."
      },
...
    },
    "schemas":{
      "DatasetPostBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Human-readable name for the dataset",
            "example": "Blood cancer database",
            "maxLength": 256
          },
          "host": {
            "type": "string",
            "description": "Url where the database is hosted at",
            "example": "http://somewhere",
            "maxLength": 256
          },
          "port": {
            "type": "number",
            "description": "Port for the host field",
            "example": 5432
          },
          "username": {
            "type": "string",
            "description": "Username for the Federated Node to access the DB",
            "example": "usernameDB"
          },
          "password": {
            "type": "string",
            "description": "Password for the Federated Node to access the DB",
            "example": "password"
          },
          "extra_connection_args": {
            "type": "string",
            "description": "Extra argument to use for connecting to the DB. Eventually this is passed to the tasks' pod as environment variable",
            "example": "TrustServerCertificate=Yes"
          },
          "type": {
            "type": "string",
            "description": "Type of DB engine",
            "oneOf": [
              {"type": "string", "example": "postgres"},
              {"type": "string", "example": "mssql"}
            ]
          },
          "catalogue": {
            "$ref": "#/components/schemas/CataloguePostBody"
          },
          "dictionaries": {
            "$ref": "#/components/schemas/DictionariesPostBody"
          }
        },
        "required": [
          "name", "username", "password", "host"
        ]
      },
...
      "TokenPostBody": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 256,
            "example": "Data Access Request 1"
          },
          "project_name": {
            "type": "string",
            "maxLength": 256,
            "example": "eu_research_2024"
          },
          "requested_by": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string",
                "maxLength": 256,
                "example": "user@email.com"
              }
            },
            "required": ["email"]
          },
          "description": {
            "type": "string",
            "maxLength": 4096,
            "description": "Project/Data access request description"
          },
          "proj_start": {
            "type": "string",
            "example": "2024-02-10"
          },
          "proj_end": {
            "type": "string",
            "example": "2025-02-25"
          },
          "dataset_id": {
            "type": "string",
            "example": "1"
          }
        },
        "required": [
          "title",
          "project_name",
          "requested_by",
          "description",
          "dataset_id",
          "proj_start",
          "proj_end"
        ]
      },
      "BeaconPostBody": {
        "type": "object",
        "properties": {
          "query":{
            "type": "string",
            "example": "SELECT col1, col2 FROM table;"
          },
          "dataset_id":{
            "type": "integer",
            "example": 1
          }
        },
        "required": [
          "query", "dataset_id"
        ]
      },
...
    }
  }
}